# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Users",
        "version": "1.0.0"
    },
    "paths": {
        "/api/health": {
            "get": {
                "tags": [
                    "API"
                ],
                "summary": "Verifica se a API está ativa",
                "operationId": "e56378d727c060d09dca88deb9289fbc",
                "responses": {
                    "200": {
                        "description": "Status OK"
                    }
                }
            }
        },
        "/api/externas": {
            "get": {
                "tags": [
                    "API"
                ],
                "summary": "Consulta dados do microserviço Node",
                "operationId": "438bdd34e74c231a170ec068f719f3ac",
                "responses": {
                    "200": {
                        "description": "Dados do microserviço retornados com sucesso"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Listar todos os usuários",
                "operationId": "7e2f5799553ec93c9c43e2a58490d447",
                "responses": {
                    "200": {
                        "description": "Lista de usuários"
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Criar um novo usuário",
                "operationId": "a1fce870c566eb5fc0cf3178ec5d2fed",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado"
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Buscar um usuário pelo ID",
                "operationId": "910a0c7cf580e1c605648d82a4e57f46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Atualizar um usuário",
                "operationId": "ddfe232db1870b82493bf8bd25596d37",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário atualizado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Remover um usuário",
                "operationId": "43dd20becb1f47444e484cbb18ce345c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Usuário removido"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "StoreUserRequest": {
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Jhon Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "Jhon_doe@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "12345678"
                    }
                },
                "type": "object"
            },
            "UpdateUserRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Jhon Doe Up"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "Jhon_doe-2025@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "novasenha123"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "API",
            "description": "Endpoints gerais da aplicação"
        },
        {
            "name": "Users",
            "description": "Gerenciamento de usuários"
        }
    ]
}